requirements.txt
asn1crypto==0.24.0
astroid
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
config==0.4.2
contextlib2==0.5.5
cryptography==2.7
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.7.1
humanize==0.5.1
idna==2.8
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jwt
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools
multitasking==0.0.9
numpy
openpyxl
packaging==19.1
pandas
pluggy==0.12.0
psutil==5.6.3
psycopg2==2.8.3
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rope==0.14.0
schema==0.7.1
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.34.2
stellar==0.4.5
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
yfinance==0.1.45
zipp==0.6.0

